Frontend: make all employee feature flags customizable
- Instead of having separate feature flags "internally" in the core and some in `lib-customizations`, make all feature flags customizable in implementations
    - One downside here is that if flags are set in a runtime-dependent way (e.g. environment-specific), modules and their dependencies cannot be dropped from builds via tree shaking
- There are basically two types of feature flags currently in eVaka:
    1. production ready features that some implementations might not want to enable
    2. Features in development (now: `experimental.*`) that can be enabled in implementations in order to test e.g. in non-production environments
- This also allows us to drop Espoo-specific assumptions about environments and their relations to URLs/domains/etc in relation to feature flags as `lib-customizations` will only export a single interface of flags and any environment specific logic is left up to the implementations
    - Espoo's environment specific configs are now located in the "espoo" theme and provided as an example for other implementations
- Also make feature flags read-only with `DeepReadOnly`
    - No need to allow overriding them in runtime
- Other frontends to follow suite
