+++ frontend/src/employee-frontend/components/ChildInformation.tsx
@@ -2,44 +2,43 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-import React, { useContext, useEffect, useMemo } from 'react'
-import styled from 'styled-components'
-import { featureFlags } from '../config'
-
-import { useTranslation } from '../state/i18n'
-import { Link, RouteComponentProps } from 'react-router-dom'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faCircle } from '@fortawesome/free-solid-svg-icons'
-import { faUsers } from 'lib-icons'
-import LocalDate from 'lib-common/local-date'
-import { Container, ContentArea } from 'lib-components/layout/Container'
-import Title from 'lib-components/atoms/Title'
-import { Gap } from 'lib-components/white-space'
-import ChildDetails from '../components/child-information/ChildDetails'
-import Assistance from '../components/child-information/Assistance'
-import FeeAlteration from '../components/child-information/FeeAlteration'
-import { UUID } from '../types'
-import { Parentship } from '../types/fridge'
-import Placements from '../components/child-information/Placements'
-import { getPersonDetails } from '../api/person'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import MessageBlocklist from 'employee-frontend/components/child-information/MessageBlocklist'
 import { Loading } from 'lib-common/api'
-import WarningLabel from '../components/common/WarningLabel'
-import { ChildContext, ChildState } from '../state/child'
-import FridgeParents from '../components/child-information/FridgeParents'
+import LocalDate from 'lib-common/local-date'
+import Title from 'lib-components/atoms/Title'
+import { Container, ContentArea } from 'lib-components/layout/Container'
+import { FixedSpaceColumn } from 'lib-components/layout/flex-helpers'
+import { Gap } from 'lib-components/white-space'
+import { featureFlags } from 'lib-customizations/employee'
+import { faUsers } from 'lib-icons'
+import React, { useContext, useEffect, useMemo } from 'react'
+import { Link, RouteComponentProps } from 'react-router-dom'
+import styled from 'styled-components'
 import { getParentshipsByChild } from '../api/parentships'
-import { UserContext } from '../state/user'
-import { TitleContext, TitleState } from '../state/title'
-import VasuAndLeops from './child-information/VasuAndLeops'
-import { getLayout, Layouts } from './layouts'
+import { getPersonDetails } from '../api/person'
+import Assistance from '../components/child-information/Assistance'
 import BackupCare from '../components/child-information/BackupCare'
 import BackupPickup from '../components/child-information/BackupPickup'
-import Guardians from '../components/child-information/Guardians'
-import FamilyContacts from '../components/child-information/FamilyContacts'
-import { requireRole } from '../utils/roles'
 import ChildApplications from '../components/child-information/ChildApplications'
-import MessageBlocklist from 'employee-frontend/components/child-information/MessageBlocklist'
+import ChildDetails from '../components/child-information/ChildDetails'
+import FamilyContacts from '../components/child-information/FamilyContacts'
+import FeeAlteration from '../components/child-information/FeeAlteration'
+import FridgeParents from '../components/child-information/FridgeParents'
+import Guardians from '../components/child-information/Guardians'
+import Placements from '../components/child-information/Placements'
+import WarningLabel from '../components/common/WarningLabel'
+import { ChildContext, ChildState } from '../state/child'
+import { useTranslation } from '../state/i18n'
+import { TitleContext, TitleState } from '../state/title'
+import { UserContext } from '../state/user'
+import { UUID } from '../types'
+import { Parentship } from '../types/fridge'
+import { requireRole } from '../utils/roles'
 import DailyServiceTimesSection from './child-information/DailyServiceTimesSection'
-import { FixedSpaceColumn } from 'lib-components/layout/flex-helpers'
+import VasuAndLeops from './child-information/VasuAndLeops'
+import { getLayout, Layouts } from './layouts'
 
 const HeaderRow = styled.div`
   display: flex;
