+++ frontend/src/employee-frontend/components/FinancePage.tsx
@@ -2,16 +2,16 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
+import Tabs from 'lib-components/molecules/Tabs'
+import { Gap } from 'lib-components/white-space'
+import { featureFlags } from 'lib-customizations/employee'
 import React, { useMemo } from 'react'
 import { Redirect, Route, Switch } from 'react-router-dom'
-import { featureFlags } from '../config'
 import { useTranslation } from '../state/i18n'
-import { RouteWithTitle } from './RouteWithTitle'
-import { Gap } from 'lib-components/white-space'
-import Tabs from 'lib-components/molecules/Tabs'
 import FeeDecisionsPage from './fee-decisions/FeeDecisionsPage'
-import VoucherValueDecisionsPage from './voucher-value-decisions/VoucherValueDecisionsPage'
 import InvoicesPage from './invoices/InvoicesPage'
+import { RouteWithTitle } from './RouteWithTitle'
+import VoucherValueDecisionsPage from './voucher-value-decisions/VoucherValueDecisionsPage'
 
 export default React.memo(function FinancePage() {
   const { i18n } = useTranslation()
@@ -23,7 +23,7 @@ export default React.memo(function FinancePage() {
         link: '/finance/fee-decisions',
         label: i18n.header.feeDecisions
       },
-      ...(featureFlags.voucherValueDecisionsPage
+      ...(featureFlags.voucher.valueDecisionsPage
         ? [
             {
               id: 'value-decisions',
@@ -53,7 +53,7 @@ export default React.memo(function FinancePage() {
           component={FeeDecisionsPage}
           title={i18n.titles.feeDecisions}
         />
-        {featureFlags.voucherValueDecisionsPage ? (
+        {featureFlags.voucher.valueDecisionsPage ? (
           <RouteWithTitle
             exact
             path="/finance/value-decisions"
