+++ frontend/src/employee-frontend/components/Header.tsx
@@ -9,7 +9,7 @@ import Title from 'lib-components/atoms/Title'
 import { FixedSpaceColumn } from 'lib-components/layout/flex-helpers'
 import { defaultMargins } from 'lib-components/white-space'
 import colors from 'lib-customizations/common'
-import { cityLogo } from 'lib-customizations/employee'
+import { cityLogo, featureFlags } from 'lib-customizations/employee'
 import { faChevronDown, faChevronUp, faSignOut } from 'lib-icons'
 import React, { useContext, useMemo, useState } from 'react'
 import {
@@ -20,7 +20,6 @@ import {
 } from 'react-router-dom'
 import styled from 'styled-components'
 import { logoutUrl } from '../api/auth'
-import { featureFlags } from '../config'
 import { isNotProduction } from '../constants'
 import { useTranslation } from '../state/i18n'
 import { UserContext } from '../state/user'
