+++ frontend/src/employee-frontend/components/login-page/Login.tsx
@@ -2,16 +2,15 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
+import Title from 'lib-components/atoms/Title'
+import { Container, ContentArea } from 'lib-components/layout/Container'
+import { Gap } from 'lib-components/white-space'
+import { featureFlags } from 'lib-customizations/employee'
 import React from 'react'
 import styled from 'styled-components'
-
-import { Container, ContentArea } from 'lib-components/layout/Container'
-import Title from 'lib-components/atoms/Title'
-import ErrorMessage from './login/ErrorMessage'
-import { useTranslation } from '../../state/i18n'
-import { Gap } from 'lib-components/white-space'
 import { getLoginUrl } from '../../api/auth'
-import { featureFlags } from '../../config'
+import { useTranslation } from '../../state/i18n'
+import ErrorMessage from './login/ErrorMessage'
 
 interface Props {
   error?: string
