+++ frontend/src/lib-customizations/espoo/featureFlags.tsx
@@ -2,12 +2,91 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-export default {
-  daycareApplication: {
-    dailyTimesEnabled: true,
-    serviceNeedOptionsEnabled: false
-  },
-  urgencyAttachmentsEnabled: true,
-  preschoolEnabled: true,
-  assistanceActionOtherEnabled: true
+import { FeatureFlags } from 'lib-customizations/types'
+
+type Env = 'staging' | 'prod'
+
+type Features = {
+  default: FeatureFlags
+} & {
+  [k in Env]: FeatureFlags
 }
+
+const env = (): Env | 'default' => {
+  if (window.location.host === 'espoonvarhaiskasvatus.fi') {
+    return 'prod'
+  }
+
+  if (window.location.host === 'staging.espoonvarhaiskasvatus.fi') {
+    return 'staging'
+  }
+
+  return 'default'
+}
+
+const features: Features = {
+  default: {
+    assistanceActionOtherEnabled: true,
+    daycareApplication: {
+      dailyTimesEnabled: true,
+      serviceNeedOptionsEnabled: false
+    },
+    evakaLogin: true,
+    financeBasicsPage: true,
+    messaging: true,
+    preschoolEnabled: true,
+    urgencyAttachmentsEnabled: true,
+    vasu: true,
+    voucher: {
+      serviceProviders: true,
+      valueDecisionsPage: true
+    },
+    experimental: {
+      ai: true
+    }
+  },
+  staging: {
+    assistanceActionOtherEnabled: true,
+    daycareApplication: {
+      dailyTimesEnabled: true,
+      serviceNeedOptionsEnabled: false
+    },
+    evakaLogin: true,
+    financeBasicsPage: true,
+    messaging: true,
+    preschoolEnabled: true,
+    urgencyAttachmentsEnabled: true,
+    vasu: false,
+    voucher: {
+      serviceProviders: true,
+      valueDecisionsPage: true
+    },
+    experimental: {
+      ai: true
+    }
+  },
+  prod: {
+    assistanceActionOtherEnabled: true,
+    daycareApplication: {
+      dailyTimesEnabled: true,
+      serviceNeedOptionsEnabled: false
+    },
+    evakaLogin: true,
+    financeBasicsPage: true,
+    messaging: true,
+    preschoolEnabled: true,
+    urgencyAttachmentsEnabled: true,
+    vasu: false,
+    voucher: {
+      serviceProviders: false,
+      valueDecisionsPage: false
+    },
+    experimental: {
+      ai: false
+    }
+  }
+}
+
+const featureFlags = features[env()]
+
+export default featureFlags
